commiting code for computation

+++ week8_3.py
@@ -1,4 +1,4 @@
-a, b = int(input("input the first number")), int(input("input the second number"))
+a, b = float(input("input the first number")), int(input("input the second number"))
 temp = a
 for i in range(b-1):
     a*=temp
